fix some things

+++ pedestrian/source/settings.h
@@ -168,7 +168,7 @@ struct Settings
 						else if (key.compare("hogFinalTreshold") == 0)
 							hogFinalTreshold = std::stod(value.c_str());
 						else if (key.compare("hogMeanshiftGrouping") == 0)
-							hogMeanshiftGrouping = std::stod(value.c_str());
+							hogMeanshiftGrouping = std::stoi(value.c_str());
 						else if (key.compare("hogGroupTreshold") == 0)
 							hogGroupTreshold = std::stoi(value.c_str());
 						else if (key.compare("hogEps") == 0)
@@ -194,7 +194,7 @@ struct Settings
 						else if (key.compare("cropHogFinalTreshold") == 0)
 							cropHogFinalTreshold = std::stod(value.c_str());
 						else if (key.compare("cropHogMeanshiftGrouping") == 0)
-							cropHogMeanshiftGrouping = std::stod(value.c_str());
+							cropHogMeanshiftGrouping = std::stoi(value.c_str());
 						else if (key.compare("cropHogGroupTreshold") == 0)
 							cropHogGroupTreshold = std::stoi(value.c_str());
 						else if (key.compare("cropHogEps") == 0)
